第42课 - 进程的初步实现（上）

+++ code/D.T.OS/screen.h
@@ -2,6 +2,8 @@
 #ifndef SCREEN_H
 #define SCREEN_H
 
+#include "type.h"
+
 #define SCREEN_WIDTH    80
 #define SCREEN_HEIGHT   25
 
@@ -16,12 +18,12 @@ typedef enum
 } PrintColor;
 
 void ClearScreen();
-int  SetPrintPos(short w, short h);
+int  SetPrintPos(byte w, byte h);
 void SetPrintColor(PrintColor c);
 int PrintChar(char c);
 int PrintString(const char* s);
 int PrintIntDec(int n);
-int PrintIntHex(unsigned int n);
+int PrintIntHex(uint n);
 
 
 #endif
